Working for weird behavior we're currently getting from the API.
Overall we need to improve how reactive we are to the API. Going to need to start assuming things in the API are failing and take action appropriately. Go through each section, ask "what happens if we don't get the section here? Should we recover or is it okay to error out? What does not having it mean?"